frontend/requirements.txt
streamlit==1.11.0
requests==2.28.1
Pillow==9.2.0
matplotlib==3.6.0